Implement q_swap() function
Using "list_del()" and "list_add(),"  implement the q_swap function.

The idea is as follows: move the "node" out of the list, then
"node->next" will be the head node for the later list. Through the
"list_add()" function, add "node" back. For "node" and "node->next"
accomplish the swap function.